Sprint3/posts_app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Smorest==0.42.0
marshmallow==3.20.1
marshmallow-enum==1.5.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pytest==7.4.2
pytest-cov==4.1.0
